+++ N3888_RefImpl/SampleApp/path_examples.cpp
@@ -30,10 +30,10 @@ namespace path_test {
 		imgSfc.paint(backBrush);
 
 		// Example code goes here.
-		pb.new_path({ 80.0, 20.0 });
-		pb.line({ 220.0, 20.0 });
-		pb.rel_line({ 60.0, 160.0 });
-		pb.rel_line({ -260.0, 0.0 });
+		pb.new_path({ 80.0F, 20.0F });
+		pb.line({ 220.0F, 20.0F });
+		pb.rel_line({ 60.0F, 160.0F });
+		pb.rel_line({ -260.0F, 0.0F });
 		pb.close_path();
 		imgSfc.stroke(foreBrush, pb, nullopt, nullopt, nullopt, aliased);
 
@@ -54,19 +54,19 @@ namespace path_test {
 		imgSfc.paint(backBrush);
 
 		// Example code goes here.
-		pb.new_path({ 20.0, 20.0 });
-		pb.rel_line({ 100.0, 0.0 });
-		pb.rel_line({ 0.0, 160.0 });
-		pb.rel_line({ -100.0, 0.0 });
-		pb.rel_line({ 0.0, -160.0 });
+		pb.new_path({ 20.0F, 20.0F });
+		pb.rel_line({ 100.0F, 0.0F });
+		pb.rel_line({ 0.0F, 160.0F });
+		pb.rel_line({ -100.0F, 0.0F });
+		pb.rel_line({ 0.0F, -160.0F });
 
-		pb.new_path({ 180.0, 20.0 });
-		pb.rel_line({ 100.0, 0.0 });
-		pb.rel_line({ 0.0, 160.0 });
-		pb.rel_line({ -100.0, 0.0 });
+		pb.new_path({ 180.0F, 20.0F });
+		pb.rel_line({ 100.0F, 0.0F });
+		pb.rel_line({ 0.0F, 160.0F });
+		pb.rel_line({ -100.0F, 0.0F });
 		pb.close_path();
 
-		imgSfc.stroke(foreBrush, pb, nullopt, stroke_props{ 10.0 }, nullopt, aliased);
+		imgSfc.stroke(foreBrush, pb, nullopt, stroke_props{ 10.0F }, nullopt, aliased);
 
 		// Example code ends.
 #ifdef _Filesystem_support_test
@@ -85,15 +85,15 @@ namespace path_test {
 		imgSfc.paint(backBrush);
 
 		// Example code goes here.
-		pb.new_path({ 20.0, 20.0 });
-		pb.rel_quadratic_curve({ 60.0, 120.0 }, { 60.0, -120.0 });
-		pb.rel_new_path({ 20.0, 0.0 });
-		pb.rel_quadratic_curve({ 60.0, 120.0 }, { 60.0, -120.0 });
+		pb.new_path({ 20.0F, 20.0F });
+		pb.rel_quadratic_curve({ 60.0F, 120.0F }, { 60.0F, -120.0F });
+		pb.rel_new_path({ 20.0F, 0.0F });
+		pb.rel_quadratic_curve({ 60.0F, 120.0F }, { 60.0F, -120.0F });
 		pb.close_path();
-		pb.new_path({ 20.0, 150.0 });
-		pb.rel_cubic_curve({ 40.0, -120.0 }, { 40.0, 120.0 * 2.0 }, { 40.0, -120.0 });
-		pb.rel_new_path({ 20.0, 0.0 });
-		pb.rel_cubic_curve({ 40.0, -120.0 }, { 40.0, 120.0 * 2.0 }, { 40.0, -120.0 });
+		pb.new_path({ 20.0F, 150.0F });
+		pb.rel_cubic_curve({ 40.0F, -120.0F }, { 40.0F, 120.0F * 2.0F }, { 40.0F, -120.0F });
+		pb.rel_new_path({ 20.0F, 0.0F });
+		pb.rel_cubic_curve({ 40.0F, -120.0F }, { 40.0F, 120.0F * 2.0F }, { 40.0F, -120.0F });
 		pb.close_path();
 		imgSfc.stroke(foreBrush, pb, nullopt, nullopt, nullopt, aliased);
 
@@ -114,36 +114,36 @@ namespace path_test {
 		imgSfc.paint(backBrush);
 
 		// Example code goes here.
-		const matrix_2d sclMtx = matrix_2d::init_scale({ 0.9, 1.1 });// .rotate(half_pi<double>);// * matrix_2d::init_rotate(half_pi<double>);
-		const vector_2d rad{ 50.0, 50.0 };
-		//auto pt = sclMtx.inverse().transform_pt({ 85.0, 100.0 }) +
-		//	point_for_angle(half_pi<double> / 2.0, rad);
-		//auto pt = sclMtx.inverse().transform_pt({ 85.0, 100.0 }) +
-		//	point_for_angle(half_pi<double> / 2.0, 50.0);
-		//auto pt = vector_2d{ 85.0, 100.0 } + 
-		//	point_for_angle(half_pi<double> / 2.0, rad);
-		auto pt = arc_start({ 85.0, 100.0 }, half_pi<double> / 2.0, { 50.0, 50.0 }, sclMtx);
+		const matrix_2d sclMtx = matrix_2d::init_scale({ 0.9F, 1.1F });// .rotate(half_pi<float>);// * matrix_2d::init_rotate(half_pi<float>);
+		const vector_2d rad{ 50.0F, 50.0F };
+		//auto pt = sclMtx.inverse().transform_pt({ 85.0F, 100.0F }) +
+		//	point_for_angle(half_pi<float> / 2.0F, rad);
+		//auto pt = sclMtx.inverse().transform_pt({ 85.0F, 100.0F }) +
+		//	point_for_angle(half_pi<float> / 2.0F, 50.0F);
+		//auto pt = vector_2d{ 85.0F, 100.0F } + 
+		//	point_for_angle(half_pi<float> / 2.0F, rad);
+		auto pt = arc_start({ 85.0F, 100.0F }, half_pi<float> / 2.0F, { 50.0F, 50.0F }, sclMtx);
 		pb.matrix(sclMtx);
 		pb.new_path(pt);
 		//pb.matrix(sclMtx);
-		pb.arc({ 50.0, 50.0 }, three_pi_over_two<double>, half_pi<double> / 2.0);
-		imgSfc.stroke(foreBrush, pb, nullopt, stroke_props{ 10.0 });// , nullopt, aliased);
+		pb.arc({ 50.0F, 50.0F }, three_pi_over_two<float>, half_pi<float> / 2.0F);
+		imgSfc.stroke(foreBrush, pb, nullopt, stroke_props{ 10.0F });// , nullopt, aliased);
 		pb.clear();
-		pb.new_path({ 130.0, 105.0 });
-		pb.rel_line({ 0.0, -10.0 });
-		pb.rel_line({ 25.0, 0.0 });
-		pb.rel_line({ 0.0, -25.0 });
-		pb.rel_line({ 10.0, 0.0 });
-		pb.rel_line({ 0.0, 25.0 });
-		pb.rel_line({ 25.0, 0.0 });
-		pb.rel_line({ 0.0, 10.0 });
-		pb.rel_line({ -25.0, 0.0 });
-		pb.rel_line({ 0.0, 25.0 });
-		pb.rel_line({ -10.0, 0.0 });
-		pb.rel_line({ 0.0, -25.0 });
+		pb.new_path({ 130.0F, 105.0F });
+		pb.rel_line({ 0.0F, -10.0F });
+		pb.rel_line({ 25.0F, 0.0F });
+		pb.rel_line({ 0.0F, -25.0F });
+		pb.rel_line({ 10.0F, 0.0F });
+		pb.rel_line({ 0.0F, 25.0F });
+		pb.rel_line({ 25.0F, 0.0F });
+		pb.rel_line({ 0.0F, 10.0F });
+		pb.rel_line({ -25.0F, 0.0F });
+		pb.rel_line({ 0.0F, 25.0F });
+		pb.rel_line({ -10.0F, 0.0F });
+		pb.rel_line({ 0.0F, -25.0F });
 		pb.close_path();
 		imgSfc.fill(foreBrush, pb);
-		pb.insert(pb.begin(), path_data::abs_matrix(matrix_2d::init_translate({ 80.0, 0.0 })));
+		pb.insert(pb.begin(), path_data::abs_matrix(matrix_2d::init_translate({ 80.0F, 0.0F })));
 		//pb.insert(pb.begin() + 2, path_data::revert_matrix());
 		imgSfc.fill(foreBrush, pb);
 		// Example code ends.
@@ -163,36 +163,36 @@ namespace path_test {
 		imgSfc.paint(backBrush);
 
 		// Example code goes here.
-		auto scl = matrix_2d::init_scale({ 0.5, 1.0 });
-		auto rot = matrix_2d::init_rotate(half_pi<double>);
-		auto trn = matrix_2d::init_translate({ 20.0, 20.0 });
-		pb.new_path({ 10.0, 150.0 });
-		pb.quadratic_curve({ 30.0, 120.0 }, { 50.0, 150.0 });
+		auto scl = matrix_2d::init_scale({ 0.5F, 1.0F });
+		auto rot = matrix_2d::init_rotate(half_pi<float>);
+		auto trn = matrix_2d::init_translate({ 20.0F, 20.0F });
+		pb.new_path({ 10.0F, 150.0F });
+		pb.quadratic_curve({ 30.0F, 120.0F }, { 50.0F, 150.0F });
 
-		//pb.new_path({ 110.0, 150.0 });
+		//pb.new_path({ 110.0F, 150.0F });
 		//pb.matrix(rot);
-		//pb.quadratic_curve({ 130.0, 120.0 }, { 150.0, 150.0 });
+		//pb.quadratic_curve({ 130.0F, 120.0F }, { 150.0F, 150.0F });
 		//pb.revert_matrix();
 
-		pb.new_path({ 110.0, 150.0 });
+		pb.new_path({ 110.0F, 150.0F });
 		pb.matrix(rot);
-		pb.quadratic_curve({ 130.0, 120.0 }, { 150.0, 150.0 });
+		pb.quadratic_curve({ 130.0F, 120.0F }, { 150.0F, 150.0F });
 		pb.revert_matrix();
 
-		pb.new_path({ 170.0, 150.0 });
+		pb.new_path({ 170.0F, 150.0F });
 		pb.matrix(rot * scl);
-		pb.rel_quadratic_curve({ 20.0, -30.0 }, { 20.0, 30.0 });
+		pb.rel_quadratic_curve({ 20.0F, -30.0F }, { 20.0F, 30.0F });
 		pb.revert_matrix();
 
-		pb.new_path({ 230.0, 150.0 });
+		pb.new_path({ 230.0F, 150.0F });
 		pb.matrix(scl * rot);
-		pb.rel_quadratic_curve({ 20.0, -30.0 }, { 20.0, 30.0 });
+		pb.rel_quadratic_curve({ 20.0F, -30.0F }, { 20.0F, 30.0F });
 		pb.revert_matrix();
 
-		pb.new_path({ 290.0, 150.0 });
+		pb.new_path({ 290.0F, 150.0F });
 		pb.matrix(rot * trn);
-		pb.rel_quadratic_curve({ 20.0, -30.0 }, { 20.0, 30.0 });
-		imgSfc.stroke(foreBrush, pb, nullopt, stroke_props{ 2.0 }, nullopt, aliased);
+		pb.rel_quadratic_curve({ 20.0F, -30.0F }, { 20.0F, 30.0F });
+		imgSfc.stroke(foreBrush, pb, nullopt, stroke_props{ 2.0F }, nullopt, aliased);
 
 		// Example code ends.
 		string fileName = "pathexample05.png";
@@ -211,22 +211,22 @@ namespace path_test {
 		path_builder<> pb{};
 		imgSfc.paint(backBrush);
 		// Example code goes here.
-		auto pt = vector_2d{ 80.0, 100.0 };
+		auto pt = vector_2d{ 80.0F, 100.0F };
 		pb.new_path(pt);
-		pb.arc({ 30.0, 30.0 }, half_pi<double>, half_pi<double>);
+		pb.arc({ 30.0F, 30.0F }, half_pi<float>, half_pi<float>);
 
-		auto rot = matrix_2d::init_rotate(half_pi<double> / 2.0);
-		pt.x(pt.x() + 50.0);
+		auto rot = matrix_2d::init_rotate(half_pi<float> / 2.0F);
+		pt.x(pt.x() + 50.0F);
 		pb.new_path(pt);
 		pb.matrix(rot);
-		pb.arc({ 30.0, 30.0 }, half_pi<double>, half_pi<double>);
+		pb.arc({ 30.0F, 30.0F }, half_pi<float>, half_pi<float>);
 
 		pb.revert_matrix();
-		rot = matrix_2d::init_rotate(half_pi<double>);
-		pt.x(pt.x() + 50.0);
+		rot = matrix_2d::init_rotate(half_pi<float>);
+		pt.x(pt.x() + 50.0F);
 		pb.new_path(pt);
-		pb.matrix(rot.translate({ 20.0, 10.0 }));
-		pb.arc({ 30.0, 30.0 }, half_pi<double>, half_pi<double>);
+		pb.matrix(rot.translate({ 20.0F, 10.0F }));
+		pb.arc({ 30.0F, 30.0F }, half_pi<float>, half_pi<float>);
 
 		imgSfc.stroke(foreBrush, pb, nullopt, nullopt, nullopt, aliased);
 		// Example code ends.
@@ -247,15 +247,15 @@ namespace path_test {
 
 		// Example code goes here.
 		brush blueBrush{ rgba_color::blue };
-		stroke_props ten{ 10.0 };
-		pb.new_path({ 30.0, 30.0 });
-		pb.rel_line({ 105.0, 0.0 });
-		pb.rel_line({ 0.0, 140.0 });
-		pb.rel_line({ -105.0, 0.0 });
+		stroke_props ten{ 10.0F };
+		pb.new_path({ 30.0F, 30.0F });
+		pb.rel_line({ 105.0F, 0.0F });
+		pb.rel_line({ 0.0F, 140.0F });
+		pb.rel_line({ -105.0F, 0.0F });
 		pb.close_path();
 		imgSfc.stroke(foreBrush, pb, nullopt, ten);
 		imgSfc.fill(blueBrush, pb);
-		pb.insert(pb.begin(), path_data::abs_matrix(matrix_2d::init_translate({ 135.0, 0.0 })));
+		pb.insert(pb.begin(), path_data::abs_matrix(matrix_2d::init_translate({ 135.0F, 0.0F })));
 		//pb.insert(pb.begin() + 2, path_data::revert_matrix());
 		imgSfc.fill(blueBrush, pb);
 		imgSfc.stroke(foreBrush, pb, nullopt, ten);
@@ -276,23 +276,23 @@ namespace path_test {
 		imgSfc.paint(backBrush);
 
 		// Example code goes here.
-		auto scl = matrix_2d::init_scale({ 0.5, 1.0 });
-		auto rot = matrix_2d::init_rotate(half_pi<double> / 2.0);
-		auto trn = matrix_2d::init_translate({ 20.0, 20.0 });
+		auto scl = matrix_2d::init_scale({ 0.5F, 1.0F });
+		auto rot = matrix_2d::init_rotate(half_pi<float> / 2.0F);
+		auto trn = matrix_2d::init_translate({ 20.0F, 20.0F });
 		pb.matrix(scl);
-		pb.new_path({ 40.0, 50.0 });
-		pb.line({ 80.0, 50.0 });
-		pb.new_path({ 40.0, 150.0 });
-		pb.rel_line({ 40.0, 0.0 });
+		pb.new_path({ 40.0F, 50.0F });
+		pb.line({ 80.0F, 50.0F });
+		pb.new_path({ 40.0F, 150.0F });
+		pb.rel_line({ 40.0F, 0.0F });
 		pb.revert_matrix();
-		pb.new_path({ 60.0, 150.0 });
-		//pb.origin({ 60.0, 150.0 });
+		pb.new_path({ 60.0F, 150.0F });
+		//pb.origin({ 60.0F, 150.0F });
 		pb.matrix(rot);
-		pb.line({ 80.0, 150.0 });
-		pb.rel_line({ 0.0, -20.0 });
+		pb.line({ 80.0F, 150.0F });
+		pb.rel_line({ 0.0F, -20.0F });
 		pb.revert_matrix();
 		//pb.revert_origin();
-		imgSfc.stroke(foreBrush, pb, nullopt, stroke_props{ 2.0 }, nullopt, aliased);
+		imgSfc.stroke(foreBrush, pb, nullopt, stroke_props{ 2.0F }, nullopt, aliased);
 		// Example code ends.
 		string fileName = "pathexample08.png";
 #ifdef _Filesystem_support_test
@@ -311,23 +311,23 @@ namespace path_test {
 		imgSfc.paint(backBrush);
 
 		// Example code goes here.
-		auto scl = matrix_2d::init_scale({ 0.5, 1.0 });
-		auto rot = matrix_2d::init_rotate(half_pi<double> / 2.0);
-		auto trn = matrix_2d::init_translate({ 20.0, 20.0 });
+		auto scl = matrix_2d::init_scale({ 0.5F, 1.0F });
+		auto rot = matrix_2d::init_rotate(half_pi<float> / 2.0F);
+		auto trn = matrix_2d::init_translate({ 20.0F, 20.0F });
 		pb.matrix(scl);
-		pb.new_path({ 40.0, 50.0 });
-		pb.line({ 80.0, 50.0 });
-		pb.new_path({ 40.0, 150.0 });
-		pb.rel_line({ 40.0, 0.0 });
+		pb.new_path({ 40.0F, 50.0F });
+		pb.line({ 80.0F, 50.0F });
+		pb.new_path({ 40.0F, 150.0F });
+		pb.rel_line({ 40.0F, 0.0F });
 		pb.revert_matrix();
-		pb.new_path({ 60.0, 150.0 });
-		//pb.origin({ 60.0, 150.0 });
+		pb.new_path({ 60.0F, 150.0F });
+		//pb.origin({ 60.0F, 150.0F });
 		pb.matrix(rot);
-		pb.line({ 80.0, 150.0 });
-		pb.rel_line({ 0.0, -20.0 });
+		pb.line({ 80.0F, 150.0F });
+		pb.rel_line({ 0.0F, -20.0F });
 		pb.revert_matrix();
 		//pb.revert_origin();
-		imgSfc.stroke(foreBrush, pb, nullopt, stroke_props{ 2.0 }, nullopt, aliased);
+		imgSfc.stroke(foreBrush, pb, nullopt, stroke_props{ 2.0F }, nullopt, aliased);
 		// Example code ends.
 		string fileName = "pathexample09.png";
 #ifdef _Filesystem_support_test
@@ -346,23 +346,23 @@ namespace path_test {
 		imgSfc.paint(backBrush);
 
 		// Example code goes here.
-		auto scl = matrix_2d::init_scale({ 0.5, 1.0 });
-		auto rot = matrix_2d::init_rotate(half_pi<double> / 2.0);
-		auto trn = matrix_2d::init_translate({ 20.0, 20.0 });
+		auto scl = matrix_2d::init_scale({ 0.5F, 1.0F });
+		auto rot = matrix_2d::init_rotate(half_pi<float> / 2.0F);
+		auto trn = matrix_2d::init_translate({ 20.0F, 20.0F });
 		pb.matrix(scl);
-		pb.new_path({ 40.0, 50.0 });
-		pb.line({ 80.0, 50.0 });
-		pb.new_path({ 40.0, 150.0 });
-		pb.rel_line({ 40.0, 0.0 });
+		pb.new_path({ 40.0F, 50.0F });
+		pb.line({ 80.0F, 50.0F });
+		pb.new_path({ 40.0F, 150.0F });
+		pb.rel_line({ 40.0F, 0.0F });
 		pb.revert_matrix();
-		pb.new_path({ 60.0, 150.0 });
-		//pb.origin({ 60.0, 150.0 });
+		pb.new_path({ 60.0F, 150.0F });
+		//pb.origin({ 60.0F, 150.0F });
 		pb.matrix(rot);
-		pb.line({ 80.0, 150.0 });
-		pb.rel_line({ 0.0, -20.0 });
+		pb.line({ 80.0F, 150.0F });
+		pb.rel_line({ 0.0F, -20.0F });
 		pb.revert_matrix();
 		//pb.revert_origin();
-		imgSfc.stroke(foreBrush, pb, nullopt, stroke_props{ 2.0 }, nullopt, aliased);
+		imgSfc.stroke(foreBrush, pb, nullopt, stroke_props{ 2.0F }, nullopt, aliased);
 		// Example code ends.
 		string fileName = "pathexample10.png";
 #ifdef _Filesystem_support_test
@@ -382,7 +382,7 @@ namespace path_test {
 //
 //		// Example code goes here.
 //
-//		imgSfc.stroke(foreBrush, pb, nullopt, stroke_props{ 2.0 }, nullopt, aliased);
+//		imgSfc.stroke(foreBrush, pb, nullopt, stroke_props{ 2.0F }, nullopt, aliased);
 //
 //		// Example code ends.
 //		string fileName = "pathexample.png";
